+++ firmware/CarbonLibrary-WifiFork.cpp
@@ -1,4 +1,4 @@
-#include "CarbonLibrary.h"
+#include "CarbonLibrary-WifiFork.h"
 
 SparkCarbon::SparkCarbon() {
     _carbonClient = NULL;
